+++ src/array/array_sync_readable_writable.rs
@@ -228,8 +228,8 @@ impl<TStorage: ?Sized + ReadableWritableStorageTraits + 'static> Array<TStorage>
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
         validate_element_size::<T>(self.data_type())?;
-        let chunk_subset_elements = crate::array::convert_to_bytes_vec(chunk_subset_elements);
-        self.store_chunk_subset_opt(chunk_indices, chunk_subset, &chunk_subset_elements, options)
+        let chunk_subset_elements = crate::array::transmute_to_bytes(chunk_subset_elements);
+        self.store_chunk_subset_opt(chunk_indices, chunk_subset, chunk_subset_elements, options)
     }
 
     #[cfg(feature = "ndarray")]
@@ -383,8 +383,8 @@ impl<TStorage: ?Sized + ReadableWritableStorageTraits + 'static> Array<TStorage>
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
         validate_element_size::<T>(self.data_type())?;
-        let subset_elements = crate::array::convert_to_bytes_vec(subset_elements);
-        self.store_array_subset_opt(array_subset, &subset_elements, options)
+        let subset_elements = crate::array::transmute_to_bytes(subset_elements);
+        self.store_array_subset_opt(array_subset, subset_elements, options)
     }
 
     #[cfg(feature = "ndarray")]